requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
bcrypt==4.3.0
billiard==4.2.1
celery==5.5.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
fastapi==0.115.12
h11==0.16.0
idna==3.10
inflection==0.5.1
kombu==5.5.3
packaging==25.0
passlib==1.7.4
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
uvicorn==0.34.2
vine==5.1.0
wcwidth==0.2.13
